Extracted MpscQueueImpl class from MpscQueue.
